+++ core/constants.php
@@ -63,8 +63,8 @@ if (PHP_SAPI !== 'cli' && !headers_sent()) {
     header('Referrer-Policy: strict-origin-when-cross-origin');
 }
 
-define('SNAPSMACK_VERSION', 'Alpha 0.7.416');
-define('SNAPSMACK_VERSION_SHORT', '0.7.416');
+define('SNAPSMACK_VERSION', 'Alpha 0.7.417');
+define('SNAPSMACK_VERSION_SHORT', '0.7.417');
 define('SNAPSMACK_VERSION_CODENAME', 'The Best Night Ever');
 
 // --- VERSION COMPARISON ---